feat: add geolocation fetching
