+++ hrms/hr/doctype/employee_checkin/employee_checkin.js
@@ -7,13 +7,17 @@ frappe.ui.form.on("Employee Checkin", {
 
 		const allow_geolocation_tracking = await frappe.db.get_single_value(
 			"HR Settings",
-			"allow_geolocation_tracking",
+			"allow_geolocation_tracking"
 		);
 
 		if (!allow_geolocation_tracking) {
 			hide_field(["fetch_geolocation", "latitude", "longitude", "geolocation"]);
 			return;
 		}
+	},
+
+	fetch_geolocation: (frm) => {
+		hrms.fetch_geolocation(frm);
 	},
 
 	add_fetch_shift_button(frm) {
@@ -38,43 +42,5 @@ frappe.ui.form.on("Employee Checkin", {
 				},
 			});
 		});
-	},
-
-	fetch_geolocation: async (frm) => {
-		if (!navigator.geolocation) {
-			frappe.msgprint({
-				message: __("Geolocation is not supported by your current browser"),
-				title: __("Geolocation Error"),
-				indicator: "red",
-			});
-			hide_field(["geolocation"]);
-			return;
-		}
-
-		frappe.dom.freeze(__("Fetching your geolocation") + "...");
-
-		navigator.geolocation.getCurrentPosition(
-			async (position) => {
-				frm.set_value("latitude", position.coords.latitude);
-				frm.set_value("longitude", position.coords.longitude);
-
-				await frm.call("set_geolocation_from_coordinates");
-				frm.dirty();
-				frappe.dom.unfreeze();
-			},
-			(error) => {
-				frappe.dom.unfreeze();
-
-				let msg = __("Unable to retrieve your location") + "<br><br>";
-				if (error) {
-					msg += __("ERROR({0}): {1}", [error.code, error.message]);
-				}
-				frappe.msgprint({
-					message: msg,
-					title: __("Geolocation Error"),
-					indicator: "red",
-				});
-			},
-		);
 	},
 });
