+++ hrms/hr/doctype/employee_checkin/employee_checkin.py
@@ -7,10 +7,8 @@ from frappe import _
 from frappe.model.document import Document
 from frappe.utils import cint, get_datetime
 
-from hrms.hr.doctype.shift_assignment.shift_assignment import (
-	get_actual_start_end_datetime_of_shift,
-)
-from hrms.hr.utils import validate_active_employee
+from hrms.hr.doctype.shift_assignment.shift_assignment import get_actual_start_end_datetime_of_shift
+from hrms.hr.utils import set_geolocation_from_coordinates, validate_active_employee
 
 
 class EmployeeCheckin(Document):
@@ -18,7 +16,7 @@ class EmployeeCheckin(Document):
 		validate_active_employee(self.employee)
 		self.validate_duplicate_log()
 		self.fetch_shift()
-		self.set_geolocation_from_coordinates()
+		set_geolocation_from_coordinates(self)
 
 	def validate_duplicate_log(self):
 		doc = frappe.db.exists(
@@ -63,28 +61,6 @@ class EmployeeCheckin(Document):
 			self.shift_actual_end = shift_actual_timings.actual_end
 			self.shift_start = shift_actual_timings.start_datetime
 			self.shift_end = shift_actual_timings.end_datetime
-
-	@frappe.whitelist()
-	def set_geolocation_from_coordinates(self):
-		if not frappe.db.get_single_value("HR Settings", "allow_geolocation_tracking"):
-			return
-
-		if not (self.latitude and self.longitude):
-			return
-
-		self.geolocation = frappe.json.dumps(
-			{
-				"type": "FeatureCollection",
-				"features": [
-					{
-						"type": "Feature",
-						"properties": {},
-						# geojson needs coordinates in reverse order: long, lat instead of lat, long
-						"geometry": {"type": "Point", "coordinates": [self.longitude, self.latitude]},
-					}
-				],
-			}
-		)
 
 
 @frappe.whitelist()
